flask_sample_app/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-logging
snowflake-connector-python
opentelemetry-instrumentation-flask
flask